revise

+++ js/script.js
@@ -10,9 +10,9 @@ function enterClicked() {
   // This function get L and W and h and calculate and show it back
 
   //input
-  const length = parseInt(document.getElementById("length").value)
-  const width = parseInt(document.getElementById("width").value)
-  const height = parseInt(document.getElementById("height").value)
+  const length = parseFloat(document.getElementById("length").value)
+  const width = parseFloat(document.getElementById("width").value)
+  const height = parseFloat(document.getElementById("height").value)
 
   // process
   const volume = (length * width * height) / 3
